+++ docs/versionhistory.rst
@@ -7,7 +7,10 @@ This library adheres to `Semantic Versioning 2.0 <http://semver.org/>`_.
 
 - Add support for ``byte``-based paths in ``connect_unix``, ``create_unix_listeners``,
   ``create_unix_datagram_socket``, and ``create_connected_unix_datagram_socket``. (PR by
-  Lura Skye.)
+  Lura Skye)
+- Fixed adjusting the total number of tokens in a ``CapacityLimiter`` on asyncio failing
+  to wake up tasks waiting to acquire the limiter in certain edge cases (fixed with help
+  from Egor Blagov)
 
 **4.1.0**